refactor: add layout components and update component paths
