+++ src/componments/Footer/index.jsx
@@ -1,14 +1,15 @@
 import React from 'react';
 import { Github, Twitter, Linkedin } from './Socials';
+import { MaxWidthContainer } from '../../layouts';
 
 function Footer() {
   return (
     <footer className="bg-lightGray-800 py-4 text-darkBlue-900 dark:bg-darkBlue-800 dark:text-lightGray-800">
-      <div className="mx-auto flex w-11/12 items-center justify-center">
+      <MaxWidthContainer compClass="flex items-center justify-center">
         <p className="flex items-center gap-2">
           Coded by Victoria | <Github /> | <Linkedin /> | <Twitter />
         </p>
-      </div>
+      </MaxWidthContainer>
     </footer>
   );
 }
