+++ src/componments/Main/index.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import Countries from '../../pages/Countries';
-import CountryPage from '../../pages/CountryPage';
+import { Countries, CountryPage } from '../../pages';
+import { MaxWidthContainer } from '../../layouts';
 
 function Main() {
   return (
     <main className="min-h-screen bg-lightGray-800 py-12 dark:bg-darkBlue-800">
-      <div className="mx-auto w-11/12 max-w-7xl">
+      <MaxWidthContainer>
         <Routes>
           <Route path="/" element={<Countries />} />
           <Route path="/country/:code" element={<CountryPage />} />
         </Routes>
-      </div>
+      </MaxWidthContainer>
     </main>
   );
 }
